+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.9.26-SNAPSHOT
 
+- introduce mongo format for java.util.Local
+
 ## 0.9.25 (2019-03-23)
 
 - artifacts for scala 2.13.0-M5 available. [95](https://github.com/sphereio/sphere-scala-libs/pull/95) and [96](https://github.com/sphereio/sphere-scala-libs/pull/96)